+++ package.json
@@ -52,7 +52,6 @@
     "prettier": "^2.4.1",
     "reflect-metadata": "^0.1.13",
     "remove-accents": "^0.4.2",
-    "string": "^3.3.3",
     "tmp": "^0.2.1",
     "toad-scheduler": "^1.6.0",
     "tsyringe": "^4.6.0",
